Move denon.ts to root of repo
